fix(signals): store scout run interval as int4 for the 30-day ceiling
PositiveSmallIntegerField (smallint, max 32767) could not hold the
documented 43200-minute (30-day) maximum the validators allow, so a
valid interval failed at DB save. Widen to PositiveIntegerField.

+++ products/signals/backend/migrations/0028_reshape_scout_config_per_scout.py
@@ -50,7 +50,7 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name="signalscoutconfig",
             name="run_interval_minutes",
-            field=models.PositiveSmallIntegerField(
+            field=models.PositiveIntegerField(
                 db_default=1440,
                 default=1440,
                 validators=[
